update: separate out script related class to Script.java and also remove BitcoinKernel.java as its not needed
